Move hash_values, sort_index, more fixtures
